+++ __tests__/integration/schema/arraysFalse.test.js
@@ -2,14 +2,14 @@ const core = require("./core");
 const { PgConnectionArgCondition } = require("graphile-build-pg");
 
 test(
-  "prints a schema with the filter plugin and the `connectionFilterLists: false` option",
+  "prints a schema with the filter plugin and the `connectionFilterArrays: false` option",
   core.test(["p"], {
     skipPlugins: [PgConnectionArgCondition],
     appendPlugins: [require("../../../index.js")],
     disableDefaultMutations: true,
     legacyRelations: "omit",
     graphileBuildOptions: {
-      connectionFilterLists: false,
+      connectionFilterArrays: false,
     },
   })
 );